Refactoring code and adding new naming conventions
Refactored code and changed tools.py -> engine_tools.py

and created tools.py that breaks previous cyclic dependencies.
