+++ retriever/engines/csvengine.py
@@ -3,8 +3,8 @@ import os
 from retriever.lib.defaults import DATA_DIR
 from retriever.lib.dummy import DummyConnection
 from retriever.lib.models import Engine
-from retriever.lib.defualts_fn import open_fw, open_csvw
-from retriever.lib.tools import sort_csv
+from retriever.lib.tools import open_fw, open_csvw
+from retriever.lib.engine_tools import sort_csv
 
 
 class engine(Engine):
