+++ retriever/engines/xmlengine.py
@@ -3,8 +3,8 @@ import os
 from retriever.lib.defaults import DATA_DIR
 from retriever.lib.dummy import DummyConnection
 from retriever.lib.models import Engine
-from retriever.lib.defualts_fn import open_fr, open_fw
-from retriever.lib.tools import xml2csv, sort_csv
+from retriever.lib.tools import open_fr, open_fw
+from retriever.lib.engine_tools import xml2csv, sort_csv
 
 
 class engine(Engine):
