+++ retriever/lib/download.py
@@ -6,7 +6,7 @@ import os
 from retriever.engines import choose_engine
 from retriever.lib.defaults import DATA_DIR, SCRIPT_WRITE_PATH
 from retriever.lib.scripts import SCRIPT_LIST
-from retriever.lib.tools import name_matches
+from retriever.lib.engine_tools import name_matches
 from retriever.lib.repository import check_for_updates
