+++ scripts/breed_bird_survey.py
@@ -16,7 +16,7 @@ from pkg_resources import parse_version
 from retriever.lib.templates import Script
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
 try:
-    from retriever.lib.scripts import open_fr, open_fw
+    from retriever.lib.tools import open_fr, open_fw
     from retriever.lib.defaults import VERSION
 except ImportError:
     from retriever import open_fr, open_fw, VERSION
