+++ scripts/breed_bird_survey_50stop.py
@@ -17,7 +17,7 @@ from retriever.lib.templates import Script
 from retriever.lib.models import Table, Cleanup, correct_invalid_value
 from pkg_resources import parse_version
 try:
-    from retriever.lib.scripts import open_fr, open_fw
+    from retriever.lib.tools import open_fr, open_fw
     from retriever.lib.defaults import VERSION
 except ImportError:
     from retriever import HOME_DIR, open_fr, open_fw, VERSION
