+++ scripts/forest_inventory_analysis.py
@@ -13,7 +13,7 @@ from retriever.lib.models import Table
 from pkg_resources import parse_version
 try:
     from retriever.lib.defaults import VERSION
-    from retriever.lib.scripts import open_fr, open_fw
+    from retriever.lib.tools import open_fr, open_fw
 except ImportError:
     from retriever import open_fr, open_fw, VERSION
