+++ test/test_retriever.py
@@ -18,13 +18,13 @@ from retriever.lib.engine import Engine
 from retriever.lib.table import Table
 from retriever.lib.templates import BasicTextTemplate
 from retriever.lib.cleanup import correct_invalid_value
-from retriever.lib.tools import getmd5
-from retriever.lib.tools import xml2csv
-from retriever.lib.tools import json2csv
-from retriever.lib.tools import sort_file
-from retriever.lib.tools import sort_csv
-from retriever.lib.tools import create_file
-from retriever.lib.tools import file_2list
+from retriever.lib.engine_tools import getmd5
+from retriever.lib.engine_tools import xml2csv
+from retriever.lib.engine_tools import json2csv
+from retriever.lib.engine_tools import sort_file
+from retriever.lib.engine_tools import sort_csv
+from retriever.lib.engine_tools import create_file
+from retriever.lib.engine_tools import file_2list
 from retriever.lib.datapackage import clean_input, is_empty
 from retriever.lib.cleanup import Cleanup
