+++ version.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 import os
 
 from retriever.lib.defaults import VERSION
-from retriever.lib.tools import get_module_version
+from retriever.lib.engine_tools import get_module_version
 
 
 def write_version_file(scripts):
